requirements.txt
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2024.7.4
click==8.1.7
colorama==0.4.6
databases==0.9.0
defusedxml==0.7.1
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
fonttools==4.53.1
fpdf2==2.7.9
getenv==0.2.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.26.4
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
port-for==0.7.2
psutil==6.0.0
psycopg==3.2.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-postgresql==6.0.1
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
rdkit==2024.3.3
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.32
alembic==1.13.2
starlette==0.37.2
typer==0.12.3
typing_extensions==4.12.2
uvicorn==0.30.3
uvloop==0.20.0
watchfiles==0.22.0
websockets==12.0
